api/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.4
fastapi==0.75.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
packaging==21.3
Pillow==9.0.1
pydantic==1.9.0
pyparsing==3.0.7
pytesseract==0.3.9
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
uvicorn==0.17.6
zipp==3.7.0
